# openapi_patched.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/scenarios": {
      "get": {
        "summary": "分页查询场景列表",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【前置条件】如有需要可配置鉴权\n【功能描述】分页查询系统或用户的场景列表数据，支持搜索功能",
        "operationId": "分页查询场景列表",
        "tags": [
          "场景模块"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "当前页码，默认 1",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页记录数，默认 10",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "场景名称或描述的搜索关键字",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenesListResponse"
                },
                "example": {
                  "code": 200,
                  "data": {
                    "items": [
                      {
                        "coverUrl": "https://example.com/cover1.jpg",
                        "description": "卧室与客厅的联合智能场景控制",
                        "id": 1,
                        "name": "智能家居控制"
                      }
                    ],
                    "total": 100
                  },
                  "message": "获取成功"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 5
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 6
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 7
          }
        },
        "security": [],
        "x-apifox-folder": "场景模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441881718-run"
      }
    },
    "/api/v1/scenarios/{id}": {
      "get": {
        "summary": "获取场景详情",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【前置条件】如有需要可配置鉴权\n【功能描述】根据场景ID获取特定场景的详细配置信息及属性数据",
        "operationId": "获取场景详情",
        "tags": [
          "场景模块"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "场景的唯一ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenesResponse"
                },
                "example": {
                  "code": 200,
                  "data": {
                    "configData": {
                      "action": "turn_on",
                      "devices": [
                        "light-1",
                        "tv-02"
                      ]
                    },
                    "coverUrl": "https://example.com/cover1.jpg",
                    "createdAt": "2026-04-01T10:00:00Z",
                    "description": "卧室与客厅的联合智能场景控制",
                    "id": 1,
                    "name": "智能家居控制",
                    "updatedAt": "2026-04-10T12:00:00Z"
                  },
                  "message": "获取成功"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 5
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 6
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 7
          }
        },
        "security": [],
        "x-apifox-folder": "场景模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441881722-run"
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "标准错误响应",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误信息"
          },
          "details": {
            "type": "object",
            "description": "详细信息",
            "x-apifox-orders": [],
            "properties": {},
            "x-apifox-ignore-properties": []
          }
        },
        "required": [
          "code",
          "message"
        ],
        "x-apifox-orders": [
          "code",
          "message",
          "details"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
      "ScenesListResponse": {
        "properties": {
          "code": {
            "description": "状态码",
            "type": "integer"
          },
          "data": {
            "description": "分页数据",
            "properties": {
              "items": {
                "description": "场景列表",
                "items": {
                  "properties": {
                    "coverUrl": {
                      "description": "场景封面图链接",
                      "type": "string"
                    },
                    "description": {
                      "description": "场景简要描述",
                      "type": "string"
                    },
                    "id": {
                      "description": "场景ID",
                      "type": "integer"
                    },
                    "name": {
                      "description": "场景名称",
                      "type": "string"
                    }
                  },
                  "type": "object",
                  "x-apifox-orders": [
                    "coverUrl",
                    "description",
                    "id",
                    "name"
                  ],
                  "x-apifox-ignore-properties": []
                },
                "type": "array"
              },
              "total": {
                "description": "总记录数",
                "type": "integer"
              }
            },
            "type": "object",
            "x-apifox-orders": [
              "items",
              "total"
            ],
            "x-apifox-ignore-properties": []
          },
          "message": {
            "description": "提示信息",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "x-apifox-orders": [
          "code",
          "data",
          "message"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
      "ScenesResponse": {
        "properties": {
          "code": {
            "description": "状态码",
            "type": "integer"
          },
          "data": {
            "description": "场景详情数据",
            "properties": {
              "configData": {
                "description": "场景的配置参数或扩展元数据",
                "type": "object",
                "x-apifox-orders": [],
                "properties": {},
                "x-apifox-ignore-properties": []
              },
              "coverUrl": {
                "description": "场景封面图链接",
                "type": "string"
              },
              "createdAt": {
                "description": "创建时间",
                "type": "string"
              },
              "description": {
                "description": "场景详尽描述",
                "type": "string"
              },
              "id": {
                "description": "场景ID",
                "type": "integer"
              },
              "name": {
                "description": "场景名称",
                "type": "string"
              },
              "updatedAt": {
                "description": "最后更新时间",
                "type": "string"
              }
            },
            "type": "object",
            "x-apifox-orders": [
              "configData",
              "coverUrl",
              "createdAt",
              "description",
              "id",
              "name",
              "updatedAt"
            ],
            "x-apifox-ignore-properties": []
          },
          "message": {
            "description": "提示信息",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "x-apifox-orders": [
          "code",
          "data",
          "message"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}